// frontend/src/app/interfaces/product.interfaces.ts
// Generated by https://quicktype.io

export interface Product {
    _id     : string;
    name    : string;
    size    : string;
    price   : number;
    qty     : number;
    tags    : string[];

}
export interface ProductCart {
    _id           : string;
    name          : string;
    size          : string;
    price         : number;
    qty           : number;
    tags          : string[];
    currentPrice  : number;
    currentQty    : number;

}